Refactor BalancingAsyncHttpClient into BalancingHttpClient
